[Attention] Fix cppsim test accumulator bitwidth
Note: Currently there is no method for optimizing the accumulator width
of both, the HLSCustomOp and the python simulation. Thus, to make the
tests pass, both must be specified manually to the maximum possible
accumulator bitwidth. Doing the MinimizeAccumulatorWidth transform would
cause the HLS and python operator behavior to diverge.