Prevent MoveOpPastFork from reordering operations on more than one axis
This is a workaround to prevent streamlining from producing a graph
which currently cannot be converted to hardware, as there is no support
for arbitrary elementwise addition with a constant parameter tensor of
arbitrary shape. Scalar and single axis elementwise addition can already
be streamlined into MultiThresholds and for multi-axis operations the
graph is technically correct, there is just no further backend support
for this, causing the creation of the streaming dataflow partition to
fail. This currently happens for (quantized) positional encodings in
transformer models.